% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{checkSchema}
\alias{checkSchema}
\title{Checks existing task file}
\usage{
checkSchema(path)
}
\arguments{
\item{path}{Path to XML file}
}
\value{
Returns \code{TRUE} if format is correct; otherwise \code{FALSE}.
}
\description{
This routine checks if an existing XML task file adheres to the schema of the
unified machine learning language.
}
\examples{
file_correct <- system.file("XML", "iris_classification.xml", package = "mlLang")
checkSchema(file_correct) # should work correctly

file_error <- system.file("XML", "iris_error.xml", package = "mlLang")
checkSchema(file_error) # should give a warning
}

