% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{executeMlTask}
\alias{executeMlTask}
\title{Executes machine learning task to train a model}
\usage{
executeMlTask(model, data, logger = cat)
}
\arguments{
\item{model}{Abstract machine learning task from loading an XML file.}

\item{data}{Data object used for training.}

\item{logger}{Specifies a function for logging. By default, the process is written to
the screen. By setting to \code{NULL}, output is disabled. When entering a filename, all
messages are written to the hard disk.}
}
\value{
Trained object of class \code{train} as return by \code{\link[caret]{train}}.
}
\description{
Function takes an abstract definition of a machine learning task and a data object as input.
It then executes the training process and returns the trained object.
}
\examples{
library(e1071)

data(iris)

# specify sample file
file <- system.file("XML", "iris_classification.xml", package = "mlLang")

# read object in unified machine learning model
task <- parseMlTask(file)

# train object with data
model <- executeMlTask(task, iris)

summary(model)
}

