% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\docType{class}
\name{MlTaskGenerator}
\alias{MlTaskGenerator}
\title{Generation of unified machine learning language}
\format{An \code{\link{R6Class}} generator object}
\usage{
MlTaskGenerator
}
\description{
R6 class for converting a machine learning task to an XML file.
}
\keyword{internal}

