% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{newMlTask}
\alias{newMlTask}
\title{Creates new machine learning task for export}
\usage{
newMlTask()
}
\value{
New object with empty task
}
\description{
Initializes a new task for machine learning and
}

