% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{class}
\name{ParsedMlTask}
\alias{ParsedMlTask}
\title{Parsed representation of unified machine learning language}
\format{An \code{\link{R6Class}} generator object}
\usage{
ParsedMlTask
}
\description{
R6 class for reading a machine learning task from  an XML file and executing it.
}
\keyword{internal}

