% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parseMlTask}
\alias{parseMlTask}
\title{Convert machine learning task from XML to R}
\usage{
parseMlTask(path)
}
\arguments{
\item{path}{Path to XML file containing the task description.}
}
\value{
An abstract representation of the task as defined by the XML file.
}
\description{
Function parses an existing machine learning task from an XML file and trains the corresponding
dataset with \code{caret}.
}
\examples{
data(iris)

# specify sample file
file <- system.file("XML", "iris_classification.xml", package = "mlLang")

# read object in unified machine learning model
task <- parseMlTask(file)

# train object with data
model <- executeMlTask(task, iris)

summary(model)
}

