% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{print.MlTaskGenerator}
\alias{print.MlTaskGenerator}
\title{Prints existing task object}
\usage{
\method{print}{MlTaskGenerator}(x, ...)
}
\arguments{
\item{x}{Object of class \code{MlTaskGenerator}.}

\item{...}{Option parameters (ignored).}
}
\value{
XML code of task object.
}
\description{
Writes XML of an existing machine learning task to the console. Output is in
unified machine learning language.
}

