% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{runMlTasks}
\alias{runMlTasks}
\title{Execute a collection of machine learning tasks}
\usage{
runMlTasks(files, data, logger = cat)
}
\arguments{
\item{files}{Collection of file names to specify XML files containing task in
unified machine learning language.}

\item{data}{Data object used for training.}

\item{logger}{Specifies a function for logging. By default, the process is written to
the screen. By setting to \code{NULL}, output is disabled. When entering a filename, all
messages are written to the hard disk.}
}
\value{
List of trained object of class \code{train} as return by
\code{\link[caret]{train}}.
}
\description{
Function executes and trains a collection of machine learning tasks.
}
\examples{
library(e1071)

data(iris)

files <- list.files(system.file("XML", package = "mlLang"),
                    pattern = "iris_classification(.*).xml",
                    full.names = TRUE)

models <- runMlTasks(files, iris)
}

