% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{scheduleMlTasks}
\alias{scheduleMlTasks}
\title{Schedules a collection of machine learning tasks for batch processing}
\usage{
scheduleMlTasks(reg, files, data, logger = NULL)
}
\arguments{
\item{reg}{Registry of batch job.}

\item{files}{Collection of file names to specify XML files containing task in
unified machine learning language.}

\item{data}{Data object used for training.}

\item{logger}{Specifies a function for logging. By default, the process is written to
the screen. By setting to \code{NULL}, output is disabled. When entering a filename, all
messages are written to the hard disk.}
}
\value{
Ids of batch jobs.
}
\description{
Function executes and trains a collection of machine learning tasks. It utilizes the
batch processing from the package \code{BatchJobs}.
}
\examples{
library(BatchJobs)

data(iris)

files <- list.files(system.file("XML", package = "mlLang"),
                    pattern = "iris_classification(.*).xml",
                    full.names = TRUE)

reg <- makeRegistry(id = "mlLang")

ids <- scheduleMlTasks(reg, files, iris)

showStatus(reg)

\dontrun{
  # retrieve result for first model once completed
  m1 <- loadResult(reg, 1)
}
}

