% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{write.MlTaskGenerator}
\alias{write}
\alias{write.MlTaskGenerator}
\title{Writes existing task object as XML to the disk}
\usage{
write.MlTaskGenerator(x, file = NULL, ...)

write(x, ...)
}
\arguments{
\item{x}{Object of class \code{MlTaskGenerator}.}

\item{file}{Filename of object.}

\item{...}{Optional arguments passed on to \code{\link[base]{write}}.}
}
\description{
Writes XML of an existing machine learning task to the disk. Output is in
unified machine learning language.
}

